refactor record utils to handler
