+++ src/coreai_opt/config/spec/compression_simulator.py
@@ -9,7 +9,6 @@ from abc import abstractmethod
 
 import torch
 import torch.nn as nn
-import torch.nn.utils.parametrize as P
 
 from coreai_opt._utils.registry_utils import ClassRegistryMixin as _ClassRegistryMixin
 
@@ -48,30 +47,3 @@ class CompressionSimulatorBase(_ClassRegistryMixin, nn.Module):
         """
         pass
 
-
-def _record_tensor_fqns_eager(model: nn.Module) -> None:
-    """Record the FQN of each parametrized tensor on the simulator compressing it."""
-    for module_name, module in model.named_modules():
-        if not P.is_parametrized(module):
-            continue
-        for param_name, parametrizations in module.parametrizations.items():
-            name = f"{module_name}.{param_name}" if module_name else param_name
-            for simulator in parametrizations:
-                if isinstance(simulator, CompressionSimulatorBase):
-                    simulator.tensor_fqn = name
-
-
-def _record_tensor_fqns_graph(model: torch.fx.GraphModule) -> None:
-    """Record the FQN of each compressed parameter on the simulator compressing it.
-
-    A simulator reading from anything other than a ``get_attr`` node keeps the
-    default name: an activation, or a weight arriving through a decompression op,
-    has no parameter to name.
-    """
-    simulators = dict(model.named_modules(remove_duplicate=False))
-    for node in model.graph.nodes:
-        if node.op != "call_module" or not node.args:
-            continue
-        simulator = simulators.get(str(node.target))
-        if isinstance(simulator, CompressionSimulatorBase) and node.args[0].op == "get_attr":
-            simulator.tensor_fqn = str(node.args[0].target)
